Serialize b-tree pages
Changes the structure of pages quite a bit, to store page internally
as bytes and only parse when needed. This is not ideal, but serves to
make things a lot easier when writing data back to disk. The sqlite
page format does not make it easy to translate back from parsed data
structures.